改行調整

+++ Assets/Scripts/Ships/Ship.cs
@@ -213,14 +213,12 @@ public partial class Ship : Things {
 
         recovery();
 
-        if(!isAlive)
-            destroyMyself();
+        if(!isAlive) destroyMyself();
 
         for(int index = 0; index < armStates.Count; index++) {
             var arm = getParts(armStates[index].partsNum);
             var hand = getHand(arm);
-            if(hand == null)
-                continue;
+            if(hand == null) continue;
 
             armStates[index].alignment = arm.setAlignment(armStates[index].alignment);
             var differenceAngle = -45 * Vector2.Angle(Vector2.left, armStates[index].alignment) / 180;
@@ -259,8 +257,7 @@ public partial class Ship : Things {
         //武装設定
         for(var index = 0; index < weaponSlots.Count; index++) {
             weaponSlots[index].partsNum = -1;
-            if(weaponSlots[index].entity == null)
-                continue;
+            if(weaponSlots[index].entity == null) continue;
             if(index < armStates.Count) {
                 getHand(getParts(armStates[index].partsNum))
                     .setWeapon(this, weaponSlots[index].entity, weaponSlots[index]);
@@ -277,8 +274,7 @@ public partial class Ship : Things {
     /// </summary>
     public void setAllAlignment(Vector2 setPosition) {
         baseAlignment = setPosition;
-        foreach(var arm in armStates)
-            arm.alignment = setPosition;
+        foreach(var arm in armStates) arm.alignment = setPosition;
     }
 
     /// <summary>
@@ -302,8 +298,7 @@ public partial class Ship : Things {
     ///燃料消費関数
     /// </summary>
     public bool reduceFuel(float reduceValue) {
-        if(!canReduceFuel(reduceValue))
-            return false;
+        if(!canReduceFuel(reduceValue)) return false;
         noReduceCount = 0;
         palamates.nowFuel = Mathf.Max(palamates.nowFuel - reduceValue, 0);
         return true;
@@ -330,18 +325,15 @@ public partial class Ship : Things {
         noDamageCount = 0;
 
         float surplusDamage = !penetration ? Mathf.Max(damage - palamates.nowBarrier, 0) : damage;
-        if(!penetration)
-            palamates.nowBarrier = Mathf.Max(palamates.nowBarrier - damage, 0);
+        if(!penetration) palamates.nowBarrier = Mathf.Max(palamates.nowBarrier - damage, 0);
 
         if(surplusDamage > 0) {
             //HPの操作
             palamates.nowArmor -= surplusDamage;
 
-            if(!continuation)
-                GetComponent<SpriteRenderer>().color = new Color(1f, 0f, 0.6f, 1);
+            if(!continuation) GetComponent<SpriteRenderer>().color = new Color(1f, 0f, 0.6f, 1);
         } else {
-            if(!continuation)
-                GetComponent<SpriteRenderer>().color = new Color(0f, 1f, 0.6f, 1);
+            if(!continuation) GetComponent<SpriteRenderer>().color = new Color(0f, 1f, 0.6f, 1);
         }
 
         setArmorBar();
@@ -360,8 +352,7 @@ public partial class Ship : Things {
         return returnVector;
     }
     public void deleteArmorBar() {
-        if(armorBar == null)
-            return;
+        if(armorBar == null) return;
         Destroy(armorBar.gameObject);
         armorBar = null;
         return;
@@ -440,12 +431,9 @@ public partial class Ship : Things {
     }
 
     protected Hand getHand(Parts target) {
-        if(target == null)
-            return null;
-        if(target.childParts == null)
-            return target.GetComponent<Hand>();
-        if(target.GetComponent<Hand>() != null)
-            return target.GetComponent<Hand>();
+        if(target == null) return null;
+        if(target.childParts == null) return target.GetComponent<Hand>();
+        if(target.GetComponent<Hand>() != null) return target.GetComponent<Hand>();
         return getHand(target.childParts);
     }
 
@@ -471,8 +459,7 @@ public partial class Ship : Things {
     public void stopAllWeapon() {
         foreach(var armstate in armStates) {
             var hand = getHand(getParts(armstate.partsNum));
-            if(hand != null)
-                hand.actionWeapon(Weapon.ActionType.NOMOTION);
+            if(hand != null) hand.actionWeapon(Weapon.ActionType.NOMOTION);
         }
     }
